requirements.txt
biopython==1.84
numpy==2.0.1
psutil==6.0.0
